sscs 9477 moved into other directory
